credit-distributor/pyproject.toml
[project]
name = "credit-distributor"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = ["skale.py==7.9dev0", "pydantic-settings==2.12.0"]

[project.optional-dependencies]
dev = ["ruff==0.14.5", "mypy==1.7.1"]

[tool.uv]
prerelease = "allow"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I"]

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
ignore_missing_imports = true
warn_return_any = true
disallow_untyped_defs = true
